digimon_data/requirements.txt
# DTCG 卡牌翻译工具依赖

# 网络请求
requests>=2.28.0

# HTML解析
beautifulsoup4>=4.11.0

# AI翻译 (可选，二选一)
google-generativeai>=0.3.0
openai>=1.0.0

# 环境变量
python-dotenv>=1.0.0
